Create Phage_high_dimension-o-matic.R
Inputs raw data as well as a file of regex. Records phage counts, dividing by total counts for that experiment to create a "PPM" value for the input (before biopanning) and output (median of 3 triplicate reads). PPM effectively represents the proportion of the total reads of that sample that represents one particular phage species.

Phage sequences are scanned for desired high-dimensional characteristics and the output value multiplied by PPM. Final values are summed and an "enrichment" score calculated which divides the output by the input.